requirements.txt
requests==2.31.0
Pillow==9.5.0
pandas==2.0.3
openpyxl==3.1.2
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
schedule==1.2.0